refactor(config): split config test files
