+++ tests/core/config/test_config.py
@@ -1,319 +1,17 @@
 import os
 import sys
 from copy import deepcopy
-from dataclasses import dataclass, field
 from pathlib import Path
-from typing import Any, Set
 
 import pytest
-import yaml
 
 from ggshield.core.config import AccountConfig, Config, InstanceConfig
 from ggshield.core.config.errors import UnknownInstanceError
-from ggshield.core.config.utils import (
-    custom_asdict,
-    get_auth_config_filepath,
-    get_global_path,
-    load_yaml,
-    replace_in_keys,
-)
+from ggshield.core.config.utils import get_auth_config_filepath, load_yaml
 from ggshield.core.constants import DEFAULT_LOCAL_CONFIG_PATH
 from ggshield.core.utils import dashboard_to_api_url
-
-
-@pytest.fixture
-def local_config_path():
-    yield "./.gitguardian"
-
-
-@pytest.fixture()
-def global_config_path():
-    yield get_global_path(".gitguardian")
-
-
-@pytest.fixture(autouse=True)
-def env_vars(monkeypatch):
-    api_key = os.getenv("TEST_GITGUARDIAN_API_KEY", "1234567890")
-    monkeypatch.setenv("GITGUARDIAN_API_URL", "https://api.gitguardian.com")
-    monkeypatch.setenv("GITGUARDIAN_API_KEY", api_key)
-
-
-def write_text(filename: str, text: str):
-    path = Path(filename)
-    path.parent.mkdir(parents=True, exist_ok=True)
-    path.write_text(text)
-
-
-def write_yaml(filename: str, data: Any):
-    path = Path(filename)
-    path.parent.mkdir(parents=True, exist_ok=True)
-    with path.open("w") as file:
-        file.write(yaml.dump(data))
-
-
-class TestUtils:
-    def test_replace_in_keys(self):
-        data = {"last-found-secrets": {"XXX"}}
-        replace_in_keys(data, "-", "_")
-        assert data == {"last_found_secrets": {"XXX"}}
-        replace_in_keys(data, "_", "-")
-        assert data == {"last-found-secrets": {"XXX"}}
-
-    def test_custom_asdict_turns_set_into_list(self):
-        """
-        GIVEN an object containing a set
-        WHEN calling test_custom_asdict() on it
-        THEN the set is turned into a list
-        """
-
-        @dataclass
-        class TestObject:
-            data: Set[str] = field(default_factory=set)
-
-        obj = TestObject(data={"a", "c", "b"})
-
-        dct = custom_asdict(obj, root=True)
-
-        assert isinstance(dct["data"], list)
-        assert sorted(dct["data"]) == ["a", "b", "c"]
-
-
-@pytest.mark.usefixtures("isolated_fs")
-class TestUserConfig:
-    def test_parsing_error(cli_fs_runner, capsys, local_config_path):
-        write_text(local_config_path, "Not a:\nyaml file.\n")
-
-        Config()
-        out, err = capsys.readouterr()
-        sys.stdout.write(out)
-        sys.stderr.write(err)
-
-        assert f"Parsing error while reading {local_config_path}:" in out
-
-    def test_display_options(self, cli_fs_runner, local_config_path):
-        write_yaml(local_config_path, {"verbose": True, "show_secrets": True})
-
-        config = Config()
-        assert config.verbose is True
-        assert config.show_secrets is True
-
-    def test_unknown_option(self, cli_fs_runner, capsys, local_config_path):
-        write_yaml(local_config_path, {"verbosity": True})
-
-        Config()
-        captured = capsys.readouterr()
-        assert "Unrecognized key in config" in captured.out
-
-    def test_display_options_inheritance(
-        self, cli_fs_runner, local_config_path, global_config_path
-    ):
-        write_yaml(
-            local_config_path,
-            {
-                "verbose": True,
-                "show_secrets": False,
-                "api_url": "https://api.gitguardian.com",
-            },
-        )
-        write_yaml(
-            global_config_path,
-            {
-                "verbose": False,
-                "show_secrets": True,
-                "api_url": "https://api.gitguardian.com2",
-            },
-        )
-
-        config = Config()
-        assert config.verbose is True
-        assert config.show_secrets is False
-        assert config.api_url == "https://api.gitguardian.com"
-
-    def test_exclude_regex(self, cli_fs_runner, local_config_path):
-        write_yaml(local_config_path, {"paths-ignore": ["/tests/"]})
-
-        config = Config()
-        assert r"/tests/" in config.paths_ignore
-
-    def test_accumulation_matches(
-        self, cli_fs_runner, local_config_path, global_config_path
-    ):
-        write_yaml(
-            local_config_path,
-            {
-                "matches_ignore": [
-                    {"name": "", "match": "one"},
-                    {"name": "", "match": "two"},
-                ]
-            },
-        )
-        write_yaml(
-            global_config_path,
-            {"matches_ignore": [{"name": "", "match": "three"}]},
-        )
-        config = Config()
-        assert config.matches_ignore == [
-            {"match": "three", "name": ""},
-            {"match": "one", "name": ""},
-            {"match": "two", "name": ""},
-        ]
-
-
-@pytest.mark.usefixtures("isolated_fs")
-class TestAuthConfig:
-    default_config = {
-        "default-token-lifetime": 2,
-        "instances": [
-            {
-                "name": "default",
-                "url": "https://dashboard.gitguardian.com",
-                "default-token-lifetime": 1,
-                "accounts": [
-                    {
-                        "workspace-id": 23,
-                        "token": "62890f237c703c92fbda8236ec2a055ac21332a46115005c976d68b900535fb5",
-                        "type": "pat",
-                        "token-name": "my_token",
-                        "expire-at": "2022-02-23T12:34:56+00:00",
-                    }
-                ],
-            },
-            {
-                "name": None,
-                "url": "https://dashboard.onprem.example.com",
-                "default-token-lifetime": 0,  # no expiry
-                "accounts": [
-                    {
-                        "workspace-id": 1,
-                        "token": "8ecffbaeedcd2f090546efeed3bc48a5f4a04a1196637aef6b3f6bbcfd58a96b",
-                        "type": "sat",
-                        "token-name": "my_other_token",
-                        "expire-at": "2022-02-24T12:34:56+00:00",
-                    }
-                ],
-            },
-        ],
-    }
-
-    def test_load(self):
-        """
-        GIVEN a default auth config
-        WHEN loading the config
-        THEN when serializing it again, it matches the data.
-        """
-        write_yaml(get_auth_config_filepath(), self.default_config)
-
-        config = Config()
-
-        assert config.instances[0].account.token_name == "my_token"
-        assert config.instances[0].default_token_lifetime == 1
-        assert config.auth_config.default_token_lifetime == 2
-
-        config_data = config.auth_config.to_dict()
-        replace_in_keys(config_data, old_char="_", new_char="-")
-        assert config_data == self.default_config
-
-    @pytest.mark.parametrize("n", [0, 2])
-    def test_no_account(self, n):
-        """
-        GIVEN an auth config with a instance with 0 or more than 1 accounts
-        WHEN loading the AuthConfig
-        THEN it raises
-        """
-        raw_config = deepcopy(self.default_config)
-        raw_config["instances"][0]["accounts"] = (
-            raw_config["instances"][0]["accounts"] * n
-        )
-        write_yaml(get_auth_config_filepath(), raw_config)
-
-        with pytest.raises(
-            AssertionError,
-            match="Each GitGuardian instance should have exactly one account",
-        ):
-            Config()
-
-    def test_invalid_format(self, capsys):
-        """
-        GIVEN an auth config file with invalid content
-        WHEN loading AuthConfig
-        THEN it raises
-        """
-        write_text(get_auth_config_filepath(), "Not a:\nyaml file.\n")
-
-        Config()
-        out, err = capsys.readouterr()
-        sys.stdout.write(out)
-        sys.stderr.write(err)
-
-        assert f"Parsing error while reading {get_auth_config_filepath()}:" in out
-
-    def test_token_not_expiring(self):
-        """
-        GIVEN an auth config file with a token never expiring
-        WHEN loading the AuthConfig
-        THEN it works
-        """
-        raw_config = deepcopy(self.default_config)
-        raw_config["instances"][0]["accounts"][0]["expire-at"] = None
-        write_yaml(get_auth_config_filepath(), raw_config)
-
-        config = Config()
-
-        assert config.instances[0].account.expire_at is None
-
-    def test_update(self):
-        """
-        GIVEN -
-        WHEN modifying the default config
-        THEN it's not persisted until .save() is called
-        """
-        config = Config()
-        config.instance = "custom"
-
-        assert Config().instance != "custom"
-
-        config.save()
-
-        assert Config().instance == "custom"
-
-    def test_load_file_not_existing(self):
-        """
-        GIVEN the auth config file not existing
-        WHEN loading the config
-        THEN it works and has the default configuration
-        """
-        config = Config()
-
-        assert config.instance_name == "https://dashboard.gitguardian.com"
-        assert config.instances == []
-
-    def test_save_file_not_existing(self):
-        """
-        GIVEN a config object and the auth config file not existing
-        WHEN saving the config
-        THEN it works and when loading the config again it has the correct values
-        """
-        config = Config()
-        try:
-            os.remove(get_auth_config_filepath())
-        except FileNotFoundError:
-            pass
-
-        config.instance = "custom"
-        config.save()
-        updated_config = Config()
-
-        assert updated_config.instance == "custom"
-
-    def test_timezone_aware_expired(self):
-        """
-        GIVEN a config with a configured instance
-        WHEN loading the config
-        THEN the instance expiration date is timezone aware
-        """
-        write_yaml(get_auth_config_filepath(), self.default_config)
-        config = Config()
-        assert config.instances[0].account.expire_at.tzinfo is not None
+from tests.conftest import write_yaml
+from tests.core.config.conftest import TEST_AUTH_CONFIG
 
 
 @pytest.mark.usefixtures("isolated_fs")
@@ -325,7 +23,7 @@ class TestConfig:
         local_instance=None,
         global_instance=None,
     ):
-        auth_config_data = deepcopy(TestAuthConfig.default_config)
+        auth_config_data = deepcopy(TEST_AUTH_CONFIG)
         for i in range(1, 6):
             config_dict = deepcopy(auth_config_data["instances"][0])
             config_dict["url"] = f"https://instance{i}.com"
